refactored namespace for getVotes
