+++ src/components/Proposals/ProposalPage/OPProposalApprovalPage/OPProposalApprovalPage.tsx
@@ -2,19 +2,22 @@ import { HStack, VStack } from "@/components/Layout/Stack";
 import ProposalDescription from "../ProposalDescription/ProposalDescription";
 import styles from "./OPProposalApprovalPage.module.scss";
 import ApprovalVotesPanel from "./ApprovalVotesPanel/ApprovalVotesPanel";
-import {
-  getUserVotesForProposal,
-  getVotesForProposal,
-  getAllForVoting,
-} from "@/app/api/votes/getVotes";
+import { getAllForVoting } from "@/app/api/votes/getVotes";
 import { getDelegate } from "@/app/api/delegates/getDelegates";
 import { Proposal } from "@/app/api/common/proposals/proposal";
 import OpManagerDeleteProposal from "../OPProposalPage/OpManagerDeleteProposal";
+import {
+  getUserVotesForProposal,
+  getVotesForProposal,
+} from "@/app/api/common/votes/getVotes";
 
 async function fetchProposalVotes(proposal_id: string, page = 1) {
   "use server";
 
-  return getVotesForProposal({ proposal_id, page });
+  return getVotesForProposal({
+    proposal_id,
+    page,
+  });
 }
 
 async function fetchAllForVoting(
